move run_test to rask_lib::testing
