+++ rask_cli/tests/integration_test.rs
@@ -2,7 +2,8 @@ use assert_cmd::Command;
 use chrono::NaiveDate;
 use diesel::prelude::*;
 use predicates::prelude::*;
-use rask_lib::{models::NewTask, schema::task};
+use rask_lib::models::NewTask;
+use rask_lib::testing::run_test;
 use regex::Regex;
 use std::{panic, str};
 
@@ -11,26 +12,6 @@ const DB_URL: &str = "postgres://postgres:password@localhost:5001/rask";
 /// Returns a connection to the database.
 fn get_db_conn() -> PgConnection {
     PgConnection::establish(DB_URL).unwrap_or_else(|_| panic!("Error connecting to {}", DB_URL))
-}
-
-/// Deletes all rows in the `task` table.
-fn delete_all_tasks(conn: &PgConnection) {
-    diesel::delete(task::table).execute(conn).unwrap();
-}
-
-/// Runs a chunk of test code in a setup/teardown block.
-/// Via https://medium.com/@ericdreichert/test-setup-and-teardown-in-rust-without-a-framework-ba32d97aa5ab .
-fn run_test<T>(test: T)
-where
-    T: FnOnce() + panic::UnwindSafe,
-{
-    let conn = get_db_conn();
-
-    let result = panic::catch_unwind(|| test());
-
-    delete_all_tasks(&conn);
-
-    assert!(result.is_ok());
 }
 
 fn get_cmd() -> Command {
@@ -94,41 +75,47 @@ fn test_list() {
 
 #[test]
 fn test_create_simple() {
-    run_test(|| {
-        let id = create_task(NewTask {
-            name: "hello there".to_string(),
-            project: None,
-            priority: None,
-            due: None,
-        });
+    run_test(
+        || {
+            let id = create_task(NewTask {
+                name: "hello there".to_string(),
+                project: None,
+                priority: None,
+                due: None,
+            });
 
-        assert_list_output_contains("Retrieved 1 tasks");
-        assert_list_output_contains("hello there");
+            assert_list_output_contains("Retrieved 1 tasks");
+            assert_list_output_contains("hello there");
 
-        assert_info_output_contains(&id, &format!("Task {}", id));
-        assert_info_output_contains(&id, "hello there");
-    })
+            assert_info_output_contains(&id, &format!("Task {}", id));
+            assert_info_output_contains(&id, "hello there");
+        },
+        get_db_conn(),
+    )
 }
 
 #[test]
 fn test_create_all_fields() {
-    run_test(|| {
-        let id = create_task(NewTask {
-            name: "clean litterbox".to_string(),
-            project: Some("frank".to_string()),
-            priority: Some("H".to_string()),
-            due: Some(NaiveDate::from_ymd(2021, 7, 31).and_hms(0, 0, 0)),
-        });
+    run_test(
+        || {
+            let id = create_task(NewTask {
+                name: "clean litterbox".to_string(),
+                project: Some("frank".to_string()),
+                priority: Some("H".to_string()),
+                due: Some(NaiveDate::from_ymd(2021, 7, 31).and_hms(0, 0, 0)),
+            });
 
-        assert_list_output_contains("Retrieved 1 tasks");
-        assert_list_output_contains("clean litterbox");
+            assert_list_output_contains("Retrieved 1 tasks");
+            assert_list_output_contains("clean litterbox");
 
-        assert_info_output_contains(&id, &format!("Task {}", id));
-        assert_info_output_contains(&id, "clean litterbox");
-        assert_info_output_contains(&id, "Project:\tfrank");
-        assert_info_output_contains(&id, "Priority:\tH");
-        assert_info_output_contains(&id, "Due:\t\t07/31/2021");
-    })
+            assert_info_output_contains(&id, &format!("Task {}", id));
+            assert_info_output_contains(&id, "clean litterbox");
+            assert_info_output_contains(&id, "Project:\tfrank");
+            assert_info_output_contains(&id, "Priority:\tH");
+            assert_info_output_contains(&id, "Due:\t\t07/31/2021");
+        },
+        get_db_conn(),
+    )
 }
 
 #[test]
