+++ rask_lib/src/lib.rs
@@ -3,3 +3,4 @@ extern crate diesel;
 
 pub mod models;
 pub mod schema;
+pub mod testing;
